refactor: added some methods fron initiative to general stats repository
